// src/db.d.ts
/**
 * This file was generated by kysely-codegen.
 * Please do not edit it manually.
 */

import type { ColumnType } from "kysely";

export type Generated<T> = T extends ColumnType<infer S, infer I, infer U>
  ? ColumnType<S, I | undefined, U>
  : ColumnType<T, T | undefined, T>;

export type Json = JsonValue;

export type JsonArray = JsonValue[];

export type JsonObject = {
  [x: string]: JsonValue | undefined;
};

export type JsonPrimitive = boolean | number | string | null;

export type JsonValue = JsonArray | JsonObject | JsonPrimitive;

export type Timestamp = ColumnType<Date, Date | string, Date | string>;

export interface Person {
  created_at: Generated<Timestamp>;
  first_name: string;
  gender: string;
  id: Generated<number>;
  last_name: string | null;
  metadata: Json | null;
  mother_id: number | null;
}

export interface Pet {
  id: Generated<number>;
  name: string;
  owner_id: number | null;
  species: string;
}

export interface DB {
  person: Person;
  pet: Pet;
}